requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
pandas>=2.0.0
plotly>=5.20.0
fpdf2>=2.7.0
